+++ projects/dxt-lossless-transform-common/src/intrinsics/color_565/mod.rs
@@ -1,0 +1,1 @@
+pub mod decorrelate;
